Set version in ZIP local header to ZIP64 when file offset is >4GB (#102053)
* ZipArchiveEntry didn't set ZIP64 in local headers for small files if their offset are > 4GB.
* Added System.IO.Compression and System.IO.Packaging tests.

---------
