+++ src/libraries/System.IO.Packaging/tests/LargeFilesTests.Net.cs
@@ -1,68 +1,34 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.IO.Compression;
+using System.Net.Mime;
 using System.Reflection;
 using Xunit;
 
-namespace System.IO.Compression.Tests;
+namespace System.IO.Packaging.Tests;
 
-[Collection(nameof(DisableParallelization))]
-public class zip_LargeFiles : ZipFileTestBase
+public partial class LargeFilesTests
 {
-    [ConditionalFact(typeof(PlatformDetection), nameof(PlatformDetection.IsSpeedOptimized), nameof(PlatformDetection.Is64BitProcess))] // don't run it on slower runtimes
-    [OuterLoop("It requires almost 12 GB of free disk space")]
-    public static void UnzipOver4GBZipFile()
-    {
-        byte[] buffer = GC.AllocateUninitializedArray<byte>(1_000_000_000); // 1 GB
-
-        string zipArchivePath = Path.Combine(Path.GetTempPath(), "over4GB.zip");
-        DirectoryInfo tempDir = Directory.CreateDirectory(Path.Combine(Path.GetTempPath(), "over4GB"));
-
-        try
-        {
-            for (byte i = 0; i < 6; i++)
-            {
-                File.WriteAllBytes(Path.Combine(tempDir.FullName, $"{i}.test"), buffer);
-            }
-
-            ZipFile.CreateFromDirectory(tempDir.FullName, zipArchivePath, CompressionLevel.NoCompression, includeBaseDirectory: false);
-
-            using ZipArchive zipArchive = ZipFile.OpenRead(zipArchivePath);
-            foreach (ZipArchiveEntry entry in zipArchive.Entries)
-            {
-                using Stream entryStream = entry.Open();
-
-                Assert.True(entryStream.CanRead);
-                Assert.Equal(buffer.Length, entryStream.Length);
-            }
-        }
-        finally
-        {
-            File.Delete(zipArchivePath);
-
-            tempDir.Delete(recursive: true);
-        }
-    }
-
     private static void FillWithHardToCompressData(byte[] buffer)
     {
         Random.Shared.NextBytes(buffer);
     }
 
     [ConditionalTheory(typeof(PlatformDetection), nameof(PlatformDetection.IsSpeedOptimized), nameof(PlatformDetection.Is64BitProcess))] // don't run it on slower runtimes
-    [OuterLoop("It requires 5~6 GB of free disk space and a lot of CPU time for compressed tests")]
     [InlineData(false)]
     [InlineData(true)]
+    [OuterLoop("It requires 5~6 GB of free disk space and a lot of CPU time for compressed tests")]
     public static void CheckZIP64VersionIsSet_ForSmallFilesAfterBigFiles(bool isCompressed)
     {
         // issue #94899
 
-        CompressionLevel compressLevel = isCompressed ? CompressionLevel.Optimal : CompressionLevel.NoCompression;
+        CompressionOption compressionOption = isCompressed ? CompressionOption.Normal : CompressionOption.NotCompressed;
         byte[] smallBuffer = GC.AllocateUninitializedArray<byte>(1000);
         byte[] largeBuffer = GC.AllocateUninitializedArray<byte>(1_000_000_000); // ~1 GB
         string zipArchivePath = Path.Combine(Path.GetTempPath(), "over4GB.zip");
-        string LargeFileName = "largefile";
-        string SmallFileName = "smallfile";
+        Uri largePartUri = PackUriHelper.CreatePartUri(new Uri("large.bin", UriKind.Relative));
+        Uri smallPartUri = PackUriHelper.CreatePartUri(new Uri("small.bin", UriKind.Relative));
         uint ZipLocalFileHeader_OffsetToVersionFromHeaderStart = 4;
         ushort Zip64Version = 45;
 
@@ -71,13 +37,14 @@ public class zip_LargeFiles : ZipFileTestBase
             using FileStream fs = File.Open(zipArchivePath, FileMode.Create, FileAccess.ReadWrite);
 
             // Create
-            using (ZipArchive archive = new(fs, ZipArchiveMode.Create, true))
+            using (Package package = Package.Open(fs, FileMode.Create, FileAccess.Write))
             {
-                ZipArchiveEntry file = archive.CreateEntry(LargeFileName, compressLevel);
+                PackagePart partLarge = package.CreatePart(largePartUri, MediaTypeNames.Application.Octet, compressionOption);
 
-                using (Stream stream = file.Open())
+                using (Stream streamLarge = partLarge.GetStream())
                 {
                     // Write 5GB of data
+
                     for (var i = 0; i < 5; i++)
                     {
                         if (isCompressed)
@@ -85,22 +52,23 @@ public class zip_LargeFiles : ZipFileTestBase
                             FillWithHardToCompressData(largeBuffer);
                         }
 
-                        stream.Write(largeBuffer);
+                        streamLarge.Write(largeBuffer);
                     }
                 }
 
-                file = archive.CreateEntry(SmallFileName, compressLevel);
+                PackagePart partSmall = package.CreatePart(smallPartUri, MediaTypeNames.Application.Octet, compressionOption);
 
-                using (Stream stream = file.Open())
+                using (Stream streamSmall = partSmall.GetStream())
                 {
-                    stream.Write(smallBuffer);
+                    streamSmall.Write(smallBuffer);
                 }
             }
+
 
             fs.Position = 0;
 
             // Validate
-            using (ZipArchive archive = new(fs, ZipArchiveMode.Read))
+            using (ZipArchive archive = new ZipArchive(fs, ZipArchiveMode.Read))
             {
                 using var reader = new BinaryReader(fs);
